Fix CX utilities/CodeGenUtil model validation bugs
From the supplied URI object in the "validateModel" method, get the
existing model resource from the UMLModeler editingDomain, instead of
creating a new resource. This change is required for the selector object
of the IBM ClientContext "com.ibm.xtools.modeler.validation.client" to
recognize the corresponding eObject resource and include the standard
constraints from this ClientContext while validating the eObject.

Also, set the option for including live constraints to 'true'. This
ensures that the constraints with evaluation mode 'live' are included in
addition to 'batch' constraints while applying the batch validation.

Closes #23