Update CodeGenUtil to enforce validation in closed models in workspace
For considering a closed resource in workspace, the second parameter of
the 'UMLModeler.getEditingDomain().getResourceSet().getResource(URI uri,
boolean loadOnDemand)' needs to be true. This method is called twice to
keep track of the 'closed/opened' status of the resource before doing
the validation. This allows us to avoid making any changes in the status
of the resource because of applying the validation. Also, precautions
have been taken to avoid any nonexistent uri passed in the validateModel
method.