Add support to upgrade existing axcioma and atcd installation (Testing).
RPM scriptlet %pre has been added to do some cleaning up before
proceeding to the installation of the features of a rpm package in the
%post scriptlet. Also, uninstallation of the features of a package is
done only if the package is really being uninstalled, not being
upgraded. The generated rpms from the travis build need to be tested
properly to see if this works.