Fix generation of unconnected ports, 1-to-many connections in a port and reorganize code
Improvements with this fix:
1. For multiple connections from a single port in a component, we used
to generate multiple connector instances. This fix changes it with the
generation of one connector instance per non-local interface port. If
any of the connected components is not deployed, RegisterNaming tag and
external connection entry tags are generated in the connector instance
associated with the port.
2. This fix provides support for the generation of unconnected ports.
This closes issue#45.
3. Also, this closes issue#50 which mentions about the generation the
same names for multiple connector instances.
4. Code for generation is reorganized. This now follows a tree
structure: while visiting a deployed component, we visit all the
non-local interface ports of the component and create all the relevant
connector artifacts that are internal to the component. Once the
generation for all the deployed components is done, we generate external
connections between the appropriate pairs of connector instances. This
reorganization incorporates more understandability and clarity in the
code generation and therefore improves the software maintainability.
5. Refactoring is done for eliminating duplicated code.