Retrieve deployment target that matches a DeploymentPart, not a CCMPart
For multiple instances of an assembly deployed as part of a root
assembly, modelElement of the same component owned by the multiple
assembly instances is not unique. Therefore, it is necessary to make
comparison with a DeploymentPart in the 'getDeploymentTarget' method,
instead of comparing with a CCMpart for retrieving the correct
DeploymentPart instance.
Also, modelElement of a deploymentPart representing an assembly can be
either a CCMComponent or a CCMPart. A correction related to this is made
in 'getUsesDelegationPartInfo' method.

This fixes issue#49.