Issue #471: refactor Xtext patches
Xtext features have a hard requirement on Log4J 1, so a Feature Patch
cannot remove this requirement.

For this reason, instead of a Feature Patch, we create new versions of
Xtext features that include the patched plugins and do not require Log4J
1.

To do this, we create a new "patches" folder with "xtext" as a
subfolder.

The feature is included in the product, but neither the update site nor
the DDK can build yet.

This commit retains the patches under papyrus-patch, but they should be
removed.

Signed-off-by: Ernesto Posse <ernesto.posse@lumenix.com>